Added color palette, download, clear and size
It is now possible to draw in different colors with different brush
thickness.
Canvas can be downloaded as a PNG. (Initially wanted to make it into a
pdf, but encountered too many problems.)
Canvas can be cleared with one button now.
The number of users displayed needs to refresh everytime for the numbers
to be correct.